requirements.txt
streamlit
langchain_community
langchain_core 
langchain_openai
langchain
python-dotenv
fastapi
flask
langserve
uvicorn